static: use ONLY pkg-config when using it
When using the static linking option on platforms that support
pkg-config, ONLY use pkg-config to get the CFLAGS and LDFLAGS so that we
pull in the libs and deps associated with the static library we've built
- and ONLY the deps necessary for it.

The main effect of this is that if someone doesn't need any sort of
remote access support at all they can completely disable libcurl,
libssh2, libssl, etc and produce a smaller/simpler binary. When done
properly, the generated pkg-config file will tell cgo everything it
needs to know.

This also prevents confusion if there is a system copy of libgit2 that
gets priority over the vendored build.

Signed-off-by: Ryan Graham <r.m.graham@gmail.com>